SCRUM-43 refactor(ToastProvider): separate logic and functional
