+++ lib/providers/ToastProvider/index.ts
@@ -1,4 +1,4 @@
 export * from './types'
 export * from './ToastContext'
 export * from './ToastProvider'
-export * from './useToast'
+export * from 'providers/ToastProvider/hooks/useToast'
